tplink protocol changes
Split tplink protocol into two scripts
Remove total accumulation. Just display daily total
